113 | Splits enabled_mod_list into monomers and multimers and loops over former
To simplify the code this splits the enabled_mod_list into a list of monomers (enabled_monomers) and multimers (enabled_multimers) and then for the enabled_monomers rather than a series of if: ...else: statements to determine what dataframes to generate and in turn concatenate the list is iterated over.

Unfortunately at the moment the tests do not pass there is one row that is different.

I've also simplified the calls to multimer_builder() to take a string argument indicating the multimer type to be built rather than having this encoded to numerical and having to decode it.