Formatting data for consolidation
Closes #145

Adds a `lowest ppm` and `Inferred Max Intensity` column to the output _only_ for rows with the absolute lowest
`delta_ppm` between observed and theoretical.

Output is now sorted by `ID` and then `delta_ppm` to aid review (although this means those that match are not at the top
of the file but in Excel).

An experimental `consolidate_matches()` function has also been introduced which attempts to perform the consolidation
step.

Tests are included and the sample data has attempted to replicate some of the example structures, including where there
are multiple matches with `delta_ppm` being equal.